pyproject.toml
[tool.poetry]
name = "autorag-live"
version = "0.1.0"
description = "A disagreement-driven, self-optimizing RAG system."
authors = ["Raj Atul Sainju <rajatsainju2025@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/rajatsainju2025/autorag-live"
repository = "https://github.com/rajatsainju2025/autorag-live"
documentation = "https://rajatsainju2025.github.io/autorag-live"
keywords = ["rag", "retrieval-augmented-generation", "nlp", "machine-learning", "search", "optimization"]
packages = [{ include = "autorag_live" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.10"
rank-bm25 = "^0.2.2"
sentence-transformers = "^2.2.2"
scipy = "^1.10.1"
typer = {extras = ["all"], version = "^0.9.0"}
streamlit = "^1.38.0"
pandas = "^2.2.0"
pyyaml = "^6.0"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
psutil = "^5.9.0"
python-dotenv = "^1.1.1"
# Optional dependencies
faiss-cpu = {version = "^1.7.4", optional = true}
qdrant-client = {version = "^1.6.0", optional = true}
elasticsearch = {version = "^8.11.0", optional = true}

[tool.poetry.extras]
all = ["faiss-cpu", "qdrant-client", "elasticsearch"]
faiss = ["faiss-cpu"]
qdrant = ["qdrant-client"]
elasticsearch = ["elasticsearch"]
search = ["elasticsearch", "qdrant-client"]
vector = ["faiss-cpu", "qdrant-client"]

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0"
pytest-cov = ">=4.0"
ruff = ">=0.1.0"
mypy = ">=1.0"
pre-commit = ">=3.0"
mkdocs = "^1.4.2"
mkdocs-material = "^9.0.0"
mkdocstrings = {extras = ["python"], version = "^0.20.0"}
click = "^8.1.3"
pytest-mock = ">=3.0"
black = ">=23.0"
isort = ">=5.0"
bandit = ">=1.7"
jupyter = ">=1.0"
ipykernel = ">=6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
autorag = "autorag_live.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=autorag_live",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=80"
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "sentence_transformers.*",
    "qdrant_client.*",
    "elasticsearch.*",
    "streamlit.*",
    "typer.*"
]
ignore_missing_imports = true
